Split PDF import file utilities
