+++ tsconfig.checkjs.json
@@ -172,6 +172,7 @@
     "js/onboarding-view-runtime.js",
     "js/onboarding-view.js",
     "js/pdf-import-ai-utils.js",
+    "js/pdf-import-file-utils.js",
     "js/pdf-import-marker-mapping.js",
     "js/pdf-import-marker-normalization.js",
     "js/pdf-import-persistence.js",
